backend/requirements.txt
Flask
Flask-CORS
Flask-SQLAlchemy
Flask-login
psycopg2-binary
python-dotenv
